Rename app_ble_gap_sec_keys
Rename app_ble_gap_sec_keys.h since it covers more than BLE security keys.
