+++ src/common/sd_rpc_impl.cpp
@@ -44,7 +44,7 @@
 #include "serialization_transport.h"
 #include "uart_boost.h"
 #include "uart_settings_boost.h"
-#include "app_ble_gap_sec_keys.h"
+#include "app_ble_gap.h"
 
 #include <cstdlib>
