+++ src/sd_api_v2/ble_gap_impl.cpp
@@ -42,7 +42,7 @@
 #include "ble_gap.h"
 #include "ble_gap_app.h"
 
-#include "app_ble_gap_sec_keys.h"
+#include "app_ble_gap.h"
 
 #include <cstdint>
