+++ src/sd_api_v2/ble_impl.cpp
@@ -40,7 +40,7 @@
 
 #include "ble.h"
 #include "ble_app.h"
-#include "app_ble_gap_sec_keys.h"
+#include "app_ble_gap.h"
 #include <adapter_internal.h>
 
 #include <cstdint>
