+++ src/sd_api_v5/ble_impl.cpp
@@ -41,7 +41,7 @@
 #include "ble.h"
 #include "ble_app.h"
 
-#include <app_ble_gap_sec_keys.h>
+#include <app_ble_gap.h>
 #include <adapter_internal.h>
 
 #include <cstdint>
